Moved BehatConfigurator to a console command.
