+++ modules/lightning_features/lightning_core/modules/lightning_dev/lightning_dev.module
@@ -4,11 +4,6 @@
  * Implements hook_modules_installed().
  */
 function lightning_dev_modules_installed(array $modules) {
-  array_walk($modules, [
-    \Drupal::service('lightning_dev.behat'),
-    'add'
-  ]);
-
   if (in_array('lightning_roles', $modules)) {
     \Drupal::service('lightning.content_roles')
       ->grantPermissions('reviewer', [
