--- modules/lightning_features/lightning_core/modules/lightning_dev/lightning_dev.services.yml
@@ -1,6 +1,0 @@
-services:
-  lightning_dev.behat:
-    class: '\Drupal\lightning_dev\BehatConfigurator'
-    arguments:
-      - '@app.root'
-      - '@module_handler'
